+++ scripts/settings.js
@@ -10,13 +10,13 @@ export function setupSettings() {
         default: true,
         type: Boolean,
     });
-    //   game.settings.register(MODULE_ID, "effect-ownership", {
-    //     name: "effect-ownership",
-    //     hint: "",
-    //     requiresReload: false,
-    //     scope: "world",
-    //     config: false,
-    //     default: true,
-    //     type: Boolean,
-    //   });
+    game.settings.register(MODULE_ID, "effect-ownership", {
+        name: `${MODULE_ID}.module-settings.effect-ownership.name`,
+        hint: `${MODULE_ID}.module-settings.effect-ownership.hint`,
+        requiresReload: false,
+        scope: "world",
+        config: true,
+        default: true,
+        type: Boolean,
+    });
 }